requirements.txt
streamlit
gTTS
playsound==1.2.2
Pillow
tensorflow